+++ src/components/IssueQueueItem.vue
@@ -7,7 +7,11 @@ const props = defineProps<{
   selected: boolean;
 }>();
 
-const nonIdFields = ['impact', 'source', 'formattedDate', 'title', 'workflowState', 'owner'];
+// Temporarily hiding 'Source' column to avoid displaying incorrect information.
+// TODO: unhide it once final issue sources are defined. [OSIDB-2424]
+//       and update the CSS for the column width in IssueQueue
+// const nonIdFields = ['impact', 'source', 'formattedDate', 'title', 'workflowState', 'owner'];
+const nonIdFields = ['impact', 'formattedDate', 'title', 'workflowState', 'owner'];
 
 const isUnembargoDateScheduledForLater = computed(
   () => DateTime.fromISO(props.issue.unembargo_dt).diffNow().milliseconds > 0,