+++ src/components/__tests__/FlawForm.spec.ts
@@ -132,7 +132,7 @@ describe('FlawForm', () => {
 
     const cvssV3Field = subject
       .findAllComponents(LabelEditable)
-      .find((component) => component.props().label === 'CVSSv3');
+      .find((component) => component.text().includes('CVSSv3') && component.text().includes('Calculator'));
     expect(cvssV3Field?.exists()).toBe(true);
 
     const cvssV3ScoreField = subject
@@ -152,7 +152,7 @@ describe('FlawForm', () => {
 
     const sourceField = subject
       .findAllComponents(LabelSelect)
-      .find((component) => component.props().label === 'Source');
+      .find((component) => component.props().label === 'CVE Source');
     expect(sourceField?.exists()).toBe(true);
 
     const statusField = subject
@@ -247,7 +247,7 @@ describe('FlawForm', () => {
 
     const sourceField = subject
       .findAllComponents(LabelSelect)
-      .find((component) => component.props().label === 'Source');
+      .find((component) => component.props().label === 'CVE Source');
     expect(sourceField?.exists()).toBe(true);
 
     const incidentStateField = subject