+++ src/types/zodFlaw.ts
@@ -111,9 +111,9 @@ export const FlawCVSSSchema = z.object({
 });
 
 export const AffectCVSSSchema = z.object({
-  affect: z.string().uuid(),
-  comment: z.string().nullable(),
-  cvss_version: z.string(),
+  affect: z.string().uuid().nullish(),
+  // comment: z.string().nullable(),
+  // cvss_version: z.string(),
   issuer: z.nativeEnum(IssuerEnum),
   score: z.number(), // $float
   uuid: z.string().uuid().nullish(), // read-only